+++ Tone/source/Noise.js
@@ -98,12 +98,12 @@ define(["Tone/core/Tone", "Tone/source/Source"], function(Tone){
 					case "brown" :
 						this._buffer = _brownNoise;
 						break;
-					default :
-						this._buffer = _whiteNoise;
+					default : 
+						throw new Error("invalid noise type: "+type)
 				}
 				//if it's playing, stop and restart it
 				if (this.state === Tone.State.Started){
-					var now = this.now() + this.bufferTime;
+					var now = this.now() + this.blockTime;
 					//remove the listener
 					this._source.onended = undefined;
 					this._stop(now);
@@ -123,7 +123,7 @@ define(["Tone/core/Tone", "Tone/source/Source"], function(Tone){
 		this._source = this.context.createBufferSource();
 		this._source.buffer = this._buffer;
 		this._source.loop = true;
-		this.connectSeries(this._source, this.output);
+		this._source.connect(this.output);
 		this.playbackRate.connect(this._source.playbackRate);
 		this._source.start(this.toSeconds(time));
 		this._source.onended = this.onended;