+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### r6
+
+* Added PitchShift and Vibrato Effect.
+* Added Timeline/TimelineState/TimelineSignal which keeps track of the state in the future
+* Clock uses requestAnimationFrame instead of ScriptProcessorNode
+* Removed `onended` event from Tone.Source
+* Refactored tests
+* Renamed some Signal methods: exponentialRampToValueNow->exponentialRampToValue, setCurrentValueNow->setRampPoint
+* LFO no longer starts at bottom of cycle. Starts at whatever phase it's set at.
+* Transport is an event emitter. triggers events on "start", "stop", "pause", and "loop". 
+* Oscillator takes a "partials" array. 
+
 ### r5
 
 * reverse buffer for Player and Sampler.
@@ -18,7 +30,7 @@
 * [new docs](http://tonejs.org/docs)
 * [updated examples](http://tonejs.org/docs)
 
-### r4 - Cool is cool
+### r4
 
 * `toFrequency` accepts notes by name (i.e. `"C4"`)
 * Envelope no longer accepts exponential scaling, only Tone.ScaledEnvelope
@@ -49,7 +61,7 @@ Or if setBpm was being invoked with a rampTime:
 	find `Tone.Transport.setBpm\((\d+)\, (\d+)\)` and replace it with `Tone.Transport.bpm.rampTo($1, $2)`.
 
 
-### r3 - Expressive Signal
+### r3
 
 Core Change:
 
@@ -84,7 +96,7 @@ Synths:
 * NoiseSynth
 
 
-### r2 - Getting Physical
+### r2
 
 * PluckSynth - Karplus-Strong Plucked String modeling synth
 * Freeverb